Initial Commit
Wrote the basic frame of some of the classes. However, I am unsure of what form of identification we should use to ID a Customer for lookup and updating.

Do tell if something that I have wrote is completely wrong or does not make sense. I learned this watching Youtube.